src/requirements.txt
beautifulsoup4==4.12.2
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pandas==2.1.0
fastapi==0.103.1
uvicorn==0.23.2
deta==1.2.0